Further cleanup of asio stream header files
 * all async ops are now collected in one header and moved to the detail namespace
 * error categories are no longer a detail, as they are visible to the user
 * more documentation in asio_stream.h
 * remove asio_includes.h helper header
