--- src/lib/tls/asio/asio_async_handshake_op.h
@@ -1,129 +1,0 @@
-/*
-* TLS Stream Helper
-* (C) 2018-2019 Jack Lloyd
-*     2018-2019 Hannes Rantzsch, Tim Oesterreich, Rene Meusel
-*
-* Botan is released under the Simplified BSD License (see license.txt)
-*/
-
-#ifndef BOTAN_ASIO_ASYNC_HANDSHAKE_OP_H_
-#define BOTAN_ASIO_ASYNC_HANDSHAKE_OP_H_
-
-#include <botan/build.h>
-
-#include <boost/version.hpp>
-#if BOOST_VERSION >= 106600
-
-#include <botan/asio_error.h>
-#include <botan/internal/asio_async_write_op.h>
-#include <botan/internal/asio_includes.h>
-
-#include <boost/asio/yield.hpp>
-
-namespace Botan {
-
-namespace TLS {
-
-template <class Handler, class Stream, class Allocator = std::allocator<void>>
-class AsyncHandshakeOperation : public AsyncBase<Handler, typename Stream::executor_type, Allocator>
-   {
-   public:
-      /**
-       * Construct and invoke an AsyncHandshakeOperation.
-       *
-       * @param handler Handler function to be called upon completion.
-       * @param stream The stream from which the data will be read
-       * @param ec Optional error code; used to report an error to the handler function.
-       */
-      template<class HandlerT>
-      AsyncHandshakeOperation(
-         HandlerT&& handler,
-         Stream& stream,
-         const boost::system::error_code& ec = {})
-         : AsyncBase<Handler, typename Stream::executor_type, Allocator>(
-              std::forward<HandlerT>(handler),
-              stream.get_executor())
-         , m_stream(stream)
-         {
-         this->operator()(ec, std::size_t(0), false);
-         }
-
-      AsyncHandshakeOperation(AsyncHandshakeOperation&&) = default;
-
-      void operator()(boost::system::error_code ec, std::size_t bytesTransferred, bool isContinuation = true)
-         {
-         reenter(this)
-            {
-            if(bytesTransferred > 0 && !ec)
-               {
-               // Provide encrypted TLS data received from the network to TLS::Channel for decryption
-               boost::asio::const_buffer read_buffer {m_stream.input_buffer().data(), bytesTransferred};
-               try
-                  {
-                  m_stream.native_handle()->received_data(
-                     static_cast<const uint8_t*>(read_buffer.data()), read_buffer.size()
-                  );
-                  }
-               catch(const TLS_Exception& e)
-                  {
-                  ec = e.type();
-                  }
-               catch(const Botan::Exception& e)
-                  {
-                  ec = e.error_type();
-                  }
-               catch(...)
-                  {
-                  ec = Botan::ErrorType::Unknown;
-                  }
-               }
-
-            if(m_stream.hasDataToSend() && !ec)
-               {
-               // Write encrypted TLS data provided by the TLS::Channel on the wire
-
-               // Note: we construct `AsyncWriteOperation` with 0 as its last parameter (`plainBytesTransferred`). This
-               // operation will eventually call `*this` as its own handler, passing the 0 back to this call operator.
-               // This is necessary because the check of `bytesTransferred > 0` assumes that `bytesTransferred` bytes
-               // were just read and are available in input_buffer for further processing.
-               AsyncWriteOperation<
-               AsyncHandshakeOperation<typename std::decay<Handler>::type, Stream, Allocator>,
-                                       Stream,
-                                       Allocator>
-                                       op{std::move(*this), m_stream, 0};
-               return;
-               }
-
-            if(!m_stream.native_handle()->is_active() && !ec)
-               {
-               // Read more encrypted TLS data from the network
-               m_stream.next_layer().async_read_some(m_stream.input_buffer(), std::move(*this));
-               return;
-               }
-
-            if(!isContinuation)
-               {
-               // Make sure the handler is not called without an intermediate initiating function.
-               // "Reading" into a zero-byte buffer will complete immediately.
-               m_ec = ec;
-               yield m_stream.next_layer().async_read_some(boost::asio::mutable_buffer(), std::move(*this));
-               ec = m_ec;
-               }
-
-            this->complete_now(ec);
-            }
-         }
-
-   private:
-      Stream&                   m_stream;
-      boost::system::error_code m_ec;
-   };
-
-}  // namespace TLS
-
-}  // namespace Botan
-
-#include <boost/asio/unyield.hpp>
-
-#endif // BOOST_VERSION
-#endif // BOTAN_ASIO_ASYNC_HANDSHAKE_OP_H_
