+++ src/lib/tls/asio/info.txt
@@ -6,15 +6,8 @@ TLS_ASIO_STREAM -> 20181218
 asio_context.h
 asio_stream.h
 asio_error.h
+asio_async_ops.h
 </header:public>
-
-<header:internal>
-asio_async_base.h
-asio_async_handshake_op.h
-asio_async_read_op.h
-asio_async_write_op.h
-asio_includes.h
-</header:internal>
 
 <requires>
 boost
